Set up some basic SHA-256 parameters and functions
Since we'll need to handle some basic things depending on the hash
algorithm, let's set up some functions and constants that handle
abstraction of things like the hash size for both SHA-1 and SHA-256
repositories.

Expose git_hash_algo_t as git_hash_algo since we'll need to expose it in
the API in several places.