oid: store the hash algorithm inside git_oid
libgit2 differs from Git in several important ways.  Most notably, the
former is a shared library that can handle multiple repositories.  As a
result, we can't assume, unlike in Git, that there is a single global
repository state that we can inspect to determine things like the hash
algorithm.

When we format a git_oid, we need to know how long it's supposed to be,
and therefore, we have to know its hash algorithm or something
equivalent to it, such as the length of the OID.  To do this, let's
create a hash_algo member and store this data there anytime we parse an
object ID.

Note that doing so necessitates changing the interface of most of our
parsing functions, including git_oid_fromstr, which we use extremely
copiously throughout the testsuite, since that function doesn't require
that the buffer be NUL terminated.  The only place we can safely infer
the algorithm is with git_oid_fromstrp, which is NUL terminated, but we
avoid doing so both to keep a consistent interface and to avoid future
problems if we add another 256-bit hash algorithm in the future.  Git
similarly tries to avoid making assumptions about the hash algorithm
based on the length when possible.

There are a few places in the code where we have a repository structure
readily available, so in those cases, we acquire the hash algorithm from
there, making small cleanups along the way.  Elsewhere, where one is not
readily available, we always use GIT_HASH_ALGO_SHA1, which we will fix
up in a later commit.

Also update git_oid_shorten for exactly the same reasons; namely, that
we need to know the length of the hash in order to shorten it
effectively.