oid: add a function to write a raw bytes from an OID
It used to be the case that we could use git_oid_cpy to copy the bytes
between a git_oid and a buffer.  Since git_oid will soon learn to have a
hash algorithm member, we'll need to provide a function that can copy
the proper number of raw bytes out, so let's do so.