tree: store full oid in entry
In the future, our git_oid structure will start having a member for the
hash algorithm, and as such, we won't want to create our tree entries by
tacking extra data onto the end of a character array.  This is
especially the case because we use an enum for our hash algorithm, which
is the same size as int and will have the same alignment requirements on
many systems, which wouldn't be satisfied by allocating at the end of an
arbitrary-size buffer.

Let's make our tree entries store a full object ID in them natively and
copy that around.